pyproject.toml
[project]
name = "vox"
version = "0.1.0"
description = "AI-powered text rewriting through macOS contextual menu"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.8.2",
    "openai>=2.16.0",
    "pyinstaller>=6.18.0",
    "pyobjc-core>=12.1",
    "pyobjc-framework-applicationservices>=12.1",
    "pyobjc-framework-cocoa>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["setuptools>=74.0", "wheel", "py2app>=0.28"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "py2app>=0.28.8",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "setuptools>=74.1.3",
]

[tool.setuptools]
packages = ["vox"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
